Reorganised files into folders
