+++ src/App.js
@@ -1,13 +1,11 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import Home from "./components/pages/Home";
+import Home from "./pages/Home";
 import "./App.css";
-import PetInfo from "./components/pages/PetInfo";
-import SubmissionForm from "./components/pages/SubmissionForm";
-import LostPets from "./components/pages/LostPets";
-import AboutUs from "./components/pages/AboutUs";
-import Login from "./components/pages/Login";
-import Registration from "./components/pages/Registration";
+import PetInfo from "./pages/PetInfo";
+import SubmissionForm from "./pages/SubmissionForm";
+import LostPets from "./pages/LostPets";
+import AboutUs from "./pages/AboutUs";
 
 
 function App() {
@@ -21,8 +19,6 @@ function App() {
           <Route path="/petInfo/:id" element={<PetInfo />} />
           <Route path="/submissionForm" element={<SubmissionForm />} />
           <Route path="/aboutUs" element={<AboutUs />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/registration" element={<Registration />} />
         </Routes>
       </Router>
