+++ src/components/filters/filter.js
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import './filter.css';
-import ButtonComponent from "../ButtonComponent";
+import ButtonComponent from "../generic/ButtonComponent";
 
 const Filter = ({ options, onClick, currentlySelected, filterMethod }) => {
     const [open, setOpen] = useState(false);
