+++ src/components/forms/SubscriptionForm.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import "./SubscriptionForm.css";
-import ButtonComponent from "../ButtonComponent";
+import ButtonComponent from "../generic/ButtonComponent";
 
 function SubscriptionForm() {
     const [email, setEmail] = useState("");
