+++ src/components/generic/pagination.css
