--- src/components/pages/Login.js
@@ -1,15 +1,0 @@
-import React from "react";
-import Navbar from "../layout/navbar";
-
-const Login = () => {
-  return (
-    <>
-      <Navbar />
-      <div style={{ paddingTop: "50px" }}>
-        <h1>This is the Login page</h1>
-      </div>
-    </>
-  );
-};
-
-export default Login;
