--- src/components/pages/Registration.js
@@ -1,15 +1,0 @@
-import React from "react";
-import Navbar from "../layout/navbar";
-
-const Registration = () => {
-  return (
-    <>
-      <Navbar />
-      <div style={{ paddingTop: "50px" }}>
-        <h1>This is the Registration page</h1>
-      </div>
-    </>
-  );
-};
-
-export default Registration;
